+++ Core/TotalTrackedTaskMinutesCalculator.cs
@@ -6,7 +6,7 @@ public class EmployeeTrackedTaskMinutes
 {
     public required long EmployeeId { get; set; }
 
-    public required decimal TrackedMinutes { get; set; }
+    public required int TrackedMinutes { get; set; }
 }
 
 public class TotalTrackedTaskMinutesCalculator
